Abstract common class for tests
